Separated DBDefinitions, added some comments
