+++ gql_granting/DBDefinitions/ProgramFormTypeModel.py
@@ -1,0 +1,22 @@
+import sqlalchemy
+from sqlalchemy import Column, String, BigInteger, Integer, DateTime, ForeignKey, Sequence, Table, Boolean, Float, DECIMAL
+from .base import BaseModel
+from .uuid import UUIDColumn, UUIDFKey
+
+class ProgramFormTypeModel(BaseModel):
+    """It encapsulates a study at university, like Cyber defence.
+
+    Args:
+        id (ID): An primary key.
+        name (str): aka Presenční
+        name_en (str): aka Present
+    """
+    __tablename__ = "acprogramforms"
+    id = UUIDColumn()
+    name = Column(String)
+    name_en = Column(String)
+
+    created = Column(DateTime, server_default=sqlalchemy.sql.func.now())
+    lastchange = Column(DateTime, server_default=sqlalchemy.sql.func.now())
+    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
+    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
