+++ gql_granting/DBDefinitions/ProgramLanguageTypeModel.py
@@ -1,0 +1,22 @@
+import sqlalchemy
+from sqlalchemy import Column, String, BigInteger, Integer, DateTime, ForeignKey, Sequence, Table, Boolean, Float, DECIMAL
+from .base import BaseModel
+from .uuid import UUIDColumn, UUIDFKey
+
+class ProgramLanguageTypeModel(BaseModel):
+    """Holds a particular language type.
+    
+    Args:
+        id (ID): An primary key.
+        name (str): aka Čeština
+        name_en (str): aka Czech
+    """
+    __tablename__ = "acprogramlanguages"
+    id = UUIDColumn()
+    name = Column(String)
+    name_en = Column(String)
+
+    created = Column(DateTime, server_default=sqlalchemy.sql.func.now())
+    lastchange = Column(DateTime, server_default=sqlalchemy.sql.func.now())
+    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
+    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
