+++ gql_granting/DBDefinitions/base.py
@@ -1,0 +1,3 @@
+from sqlalchemy.orm import declarative_base
+
+BaseModel = declarative_base()
