+++ gql_granting/DBDefinitions/uuid.py
@@ -1,0 +1,27 @@
+import uuid
+from sqlalchemy import (
+    Column,
+    String
+)
+
+def newUuidAsString():
+    return f"{uuid.uuid1()}"
+
+
+def UUIDColumn(name=None):
+    if name is None:
+        return Column(String, primary_key=True, unique=True, default=newUuidAsString)
+    else:
+        return Column(
+            name, String, primary_key=True, unique=True, default=newUuidAsString
+        )
+
+def UUIDFKey(*, ForeignKey=None, nullable=False):
+    if ForeignKey is None:
+        return Column(
+            String, index=True, nullable=nullable
+        )
+    else:
+        return Column(
+            ForeignKey, index=True, nullable=nullable
+        )
